fix: aplica correcoes review
